engine, ecs: fixes fixes fixes
- acquire data from event queue
because one can send an event during event processing
and if this event depends on new template creation with
queues recalculation it must be processed next frame
-move data from INIT comps instead of copying them
-use memcpy if component is trivial (also looks like I've forgot
to use trivialMovement anywhere in the code base .____.)
-use placement new in the template component initialization
instead of memcpy for nonboxed types. Because it not expected
to be trivial (i.e. it might be a handler with destructor)
-added handler type manager basically for gapi resources
and allow to specify custom name in ECS type registration